added contrast comparison sims
Updated the source program to include a function that simulates both the experimental and comparative effectiveness contrasts for clinical utility. Also added the program simulation.shih.com.R to call the simulation function, and we added a function in analysis.func.R to analyze the results. Results were added in the Results folder.